Commit with modules and babel
