+++ src/index.html
@@ -98,6 +98,6 @@
     </div>
 
   </div>
-  <script src="script.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 </html>
