+++ webpack.config.js
@@ -6,5 +6,14 @@ module.exports = {
     output: {
         path: path.resolve(__dirname, 'dist'),
         filename: 'main.js'
-    }
+    },
+
+    module: {
+        rules: [{ // тут описываются правила
+            test: /\.js$/, // регулярное выражение, которое ищет все js файлы
+            use: { loader: "babel-loader" }, // весь JS обрабатывается пакетом babel-loader
+            exclude: /node_modules/ // исключает папку node_modules
+                }
+            ]
+        }
 }
